% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay-class.R
\docType{class}
\name{Assay-class}
\alias{Assay-class}
\alias{Assay}
\alias{results}
\alias{plot,Assay,missing-method}
\alias{results,Assay-method}
\title{Assay constructor}
\usage{
Assay(df, model = function(x) stats::lm(value ~ conc, data = x))

\S4method{plot}{Assay,missing}(x, samples = TRUE)

\S4method{results}{Assay}(x, standards = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} containing one row of values for each sample.}

\item{model}{A \code{function} describing the relationship between assay values and
standards.}

\item{x}{An Assay object.}

\item{samples}{Should interpolated sample values be included on the standard curve?}

\item{standards}{Should the standards be returned with the results?}
}
\description{
\code{Assay} is an S4 container for interpolating sample values from a standard
curve. When provided a data frame (\code{df}) containing one row for each
measurement and a \code{model} function for the relationship, unknown values are
interpolated from the standard curve. The standard curve can be evaluated using
\code{plot} and the interpolated values are returned as a data frame using \code{results}.
}
\examples{
(x <- Assay(assay, model = \(x) stats::lm(value ~ conc, data = x)))

plot(x, samples = TRUE)

results(x, standards = TRUE)

}
