% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assay}
\alias{assay}
\title{Assay Data for Interpolation}
\format{
\describe{
\item{id}{sample ID}
\item{sample}{sample descriptors}
\item{conc}{concentration of standards}
\item{value}{assay result}
}
}
\usage{
assay
}
\description{
A toy dataset formatted for interpolation of sample concentrations from a
standard curve.
}
\keyword{datasets}
