% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Values Not In}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{The values to be matched.}

\item{table}{The values to matched against.}
}
\value{
A logical vector indicating if a match was not located for each
element of \code{x}.
}
\description{
This returns the negation of \code{\%in\%}.
}
\examples{
1 \%nin\% 1:4
}
