% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{impute_qrilc}
\alias{impute_qrilc}
\title{Impute missing values using QRILC}
\usage{
impute_qrilc(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A SummarizedExperiment object without missing assay values.
}
\description{
This method is designed to impute missing values from left-censored data, such
as targeted metabolomics data where missingness is more likely related to
low signal intensities.
}
\examples{
se <- impute_qrilc(se)

}
