% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{impute_rf}
\alias{impute_rf}
\title{Impute missing values using random forest}
\usage{
impute_rf(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A SummarizedExperiment object without missing assay values.
}
\description{
This function will impute missing values in a SummarizedExperiment assay slot
using random forest. The seed is set inside the function so that it will return
the same replacement values each time it is called.
}
\examples{
se <- impute_rf(se)

}
