% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{norm_pqn}
\alias{norm_pqn}
\title{Normalize using probabilistic quotient normalization}
\usage{
norm_pqn(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
A SummarizedExperiment object with normalized assay values.
}
\description{
This function will perform a probabilistic quotient normalization on the assay
values in a SummarizedExperiment object. If the \code{rowData} contains a column
named \code{reference}, these values will be used for normalization, otherwise, the
mean feature value across all samples will be used.
}
\examples{
se <- norm_pqn(se)

}
