% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize experimental means across batches}
\usage{
normalize(.data, .value, ...)
}
\arguments{
\item{.data}{A data frame}

\item{.value}{Column name of experimental values}

\item{...}{Column name(s) of grouping variables}
}
\value{
A data frame
}
\description{
This function will correct sample values by the difference between batch means
and the grand mean of the measurement value.
}
\examples{
df <- assay[!is.na(assay$sample), ]
normalize(df, value, sample)

}
