% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics.R
\name{rm_missing}
\alias{rm_missing}
\title{Filter missing metabolites}
\usage{
rm_missing(x, group = NULL, frac = 0.2)
}
\arguments{
\item{x}{A tibble formatted for the \code{as_SummarizedExperiment} method. The column
containing assay data should be \code{a.NAME}. The columns containing feature and
sample IDs should be \code{f.id} and \code{s.id}, respectively. The remaining columns
should be tagged as either feature ("\code{f.NAME}") or sample ("\code{s.NAME}") values.
This format is returned by the \code{as_tibble} method.}

\item{group}{A character vector of column names defining experiment groups. The
default value will remove features where > 20\% values are missing.}

\item{frac}{The maximum fraction of missing values tolerated for a feature.}
}
\value{
A tibble.
}
\description{
Remove features with missingness greater than a defined fraction (0.2 is the
default). The \code{group} argument may be used to define experimental groups, in
which case, features will only be removed if all groups individually have
missingness greater than the defined fraction. If no experimental groups are
defined, then features are removed based on the missingness across all samples.
}
\examples{
# add missing values to feature 1
assay <- SummarizedExperiment::assay(se)
assay[(1 + 200 * c(0, 2, 3))] <- NA
SummarizedExperiment::assay(se) <- assay
assay

# remove based on missingness across all samples
tibble::as_tibble(se) |> rm_missing(frac = 0.4) # f.id == 1 removed

# remove based on missingness within groups
tibble::as_tibble(se) |> rm_missing(group = "s.treatment", frac = 0.4) # f.id == 1 preserved

}
