% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se}
\alias{se}
\title{An example SummarizedExperiment}
\format{
An object of class \code{SummarizedExperiment} with 200 rows and 6 columns.
}
\usage{
se
}
\description{
A toy data set.
}
\keyword{datasets}
