% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tackle-package.R
\docType{package}
\name{tackle-package}
\alias{tackle}
\alias{tackle-package}
\title{tackle: Personal Utility Functions}
\description{
A collection of personal utilty functions for data analysis and bioinformatics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmoldham/tackle}
  \item Report bugs at \url{https://github.com/wmoldham/tackle/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
