% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R, R/as_SummarizedExperiment.R
\name{as_tibble.SummarizedExperiment}
\alias{as_tibble.SummarizedExperiment}
\alias{as_SummarizedExperiment}
\alias{as_SummarizedExperiment.tbl_df}
\title{Convert between SummarizedExperiments and Tibbles}
\usage{
\method{as_tibble}{SummarizedExperiment}(se, a.name = "value")

as_SummarizedExperiment(tbl)

\method{as_SummarizedExperiment}{tbl_df}(tbl)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object.}

\item{a.name}{The name for the column of assay data (\emph{e.g.}, "counts").}

\item{tbl}{A \code{tbl_df} object with appropriately annotated column names.}
}
\value{
A \code{tbl} containing one row for each combination of sample and feature.

A \code{SummarizedExperiment}
}
\description{
\code{as_tibble} will convert a SummarizedExperiment into a \code{tbl_df}. To keep track
of feature, sample, and assay data, the column names are annotated.

\code{as_SummarizedExperiment} will convert a \code{tbl} into a SummarizedExperiment based
on column name annotations. The column containing assay data should be \code{a.NAME}.
The columns containing feature and sample IDs should be \code{f.id} and \code{s.id},
respectively. The remaining columns should be tagged as either feature ("\code{f.NAME}")
or sample ("\code{s.NAME}") values. This format is returned by the \code{as_tibble} method.
Since a SummarizedExperiment requires one row for each feature, sample-specific
differences in features (\emph{e.g.}, retention times) are summarized as the mean
and range across all samples.
}
\examples{
tbl <- tibble::as_tibble(se, a.name = "counts")

as_SummarizedExperiment(tbl)

}
