% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assay}
\alias{assay}
\title{Assay data for interpolation}
\format{
An object of class \code{data.frame} with 9 rows and 4 columns.
}
\usage{
assay
}
\description{
A toy data set formatted for interpolation of sample concentrations from a
standard curve.
}
\details{
\describe{
\item{id}{sample ID}
\item{sample}{sample descriptors}
\item{conc}{concentration of standards}
\item{value}{assay result}
}
}
\keyword{datasets}
