% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_se.R
\name{tbl_se}
\alias{tbl_se}
\alias{as_tibble.SummarizedExperiment}
\alias{as_SummarizedExperiment.tbl_se}
\alias{as_SummarizedExperiment}
\title{Convert between SummarizedExperiments and Tibbles}
\usage{
tbl_se(tbl, assay, features, feature_data = NULL, samples, sample_data = NULL)

\method{as_tibble}{SummarizedExperiment}(x, assay_name)

as_SummarizedExperiment(tbl)

\method{as_SummarizedExperiment}{tbl_se}(tbl)
}
\arguments{
\item{tbl}{A \code{tbl_se} object generated from \code{tbl_se} or \code{as_tibble}.}

\item{assay}{The column name of the assay data.}

\item{features}{The column name of the feature data or IDs.}

\item{feature_data}{The column name(s) of additional feature information.}

\item{samples}{The column name of the sample data or IDs.}

\item{sample_data}{The column name(s) of additional sample information.}

\item{x}{A \code{SummarizedExperiment} object.}

\item{assay_name}{The name for the column of assay data (\emph{e.g.}, "counts").}
}
\description{
A \code{tbl_se} identifies the assay, feature, and sample data required to convert
a Tibble into a SummarizedExperiment.

\code{as_tibble} will convert a SummarizedExperiment into a \code{tbl_se}.

\code{as_SummarizedExperiment} will convert a \code{tbl_se} into a SummarizedExperiment.
Since a SummarizedExperiment requires one row for each feature, sample-specific
differences in features (\emph{i.e.}, retention times) are summarized as the mean
and range across all samples.
}
\examples{
x <- tibble::as_tibble(se, assay_name = "counts")

as_SummarizedExperiment(x)

}
